Added wallet to Customer

+++ src/main/java/com/example/demo/Model/Payment/LoyaltyPointsDecorator.java
@@ -7,7 +7,7 @@ public class LoyaltyPointsDecorator extends PaymentDecorator {
     }
 
     @Override
-    public int processPayment() {
+    public float processPayment() {
         float loyaltyDiscountedAmount = decoratedPayment.getAmount() * 0.90f; // Apply 10% discount
         decoratedPayment.setAmount(loyaltyDiscountedAmount);
         System.out.println("10% Loyalty Discount Applied.");
